entity 수정(Product, Item)

+++ src/main/java/com/traceprice/takeoffer/entity/Item.java
@@ -19,7 +19,7 @@ public class Item {
     private String pname;
     private Integer fixed_price;
     private String item_img;
-    private Integer item_number;
+    private Long item_number;
 
     @ManyToOne
     @JoinColumn(name = "p_id")
